+++ driver.go
@@ -6,12 +6,13 @@ import (
 	"path"
 	"time"
 
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/metadata"
+
 	"github.com/yandex-cloud/ydb-go-sdk/v2/api/protos/Ydb"
 	"github.com/yandex-cloud/ydb-go-sdk/v2/api/protos/Ydb_Operations"
 	"github.com/yandex-cloud/ydb-go-sdk/v2/internal"
 	"github.com/yandex-cloud/ydb-go-sdk/v2/timeutil"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/metadata"
 )
 
 type driver struct {